feat: Fix MyPy strict type errors
This commit fixes a number of MyPy strict type errors and failing tests.

Changes include:
- Installing `types-requests` and `types-pyyaml` to provide missing type stubs.
- Adding `isinstance` checks to handle union types returned by the Alpaca SDK.
- Correcting the `GetOrdersRequest` status from an enum to a string.
- Fixing several failing tests by using `MagicMock(spec=Order)` and correcting mock object attributes.
- Defining the `RejectedSignal` Pydantic model.

While many issues have been resolved, some tests in `test_trade_archival.py` are still failing. The root cause appears to be related to the mocking of `Order` objects, but a definitive solution has not yet been found.

+++ tests/analysis/test_structural.py
@@ -287,11 +287,11 @@ class TestPerformance:
         lows = (base - np.abs(np.random.randn(n)) * 2).astype(np.float64)
 
         # Warm-up run (first call may include some overhead)
-        _zigzag_core(highs[:1000], lows[:1000], 0.05)
+        _zigzag_core(highs[:1000].astype(np.float64), lows[:1000].astype(np.float64), 0.05)
 
         # Timed run
         start = time.perf_counter()
-        result = _zigzag_core(highs, lows, 0.05)
+        result = _zigzag_core(highs.astype(np.float64), lows.astype(np.float64), 0.05)
         elapsed = time.perf_counter() - start
 
         # Should complete in under 50ms (relaxed from 5ms for CI variance)
